+++ run.sh
@@ -1,6 +1,6 @@
 category=services
 service=history-parse
-version=0.1.8
+version=1.0.0
 
 cd ./../../automation_scripts