+++ src/history-parse/Aggregate/Entities/ParsedMatch.cs
@@ -1,13 +1,19 @@
 ﻿using System;
-using CSGOStats.Services.Core.Handling.Entities;
+using CSGOStats.Infrastructure.Core.Data.Entities;
+using NodaTime;
 
 namespace CSGOStats.Services.HistoryParse.Aggregate.Entities
 {
     public class ParsedMatch : AggregateRoot
     {
-        public ParsedMatch(Guid id, long version)
+        public OffsetDateTime CreatedAt { get; private set; }
+
+        public ParsedMatch() { /* EF */ }
+
+        public ParsedMatch(Guid id, long version, OffsetDateTime createdAt)
             : base(id, version)
         {
+            CreatedAt = createdAt;
         }
     }
 }