+++ src/history-parse/Aggregate/Factories/ParsedMatchFactory.cs
@@ -1,11 +1,12 @@
 ﻿using System;
-using CSGOStats.Services.Core.Handling.Entities;
+using CSGOStats.Infrastructure.Core.Data.Entities;
+using CSGOStats.Infrastructure.Core.Extensions;
 using CSGOStats.Services.HistoryParse.Aggregate.Entities;
 
 namespace CSGOStats.Services.HistoryParse.Aggregate.Factories
 {
     public class ParsedMatchFactory : IEntityFactory<ParsedMatch, Guid>
     {
-        public ParsedMatch CreateEmpty(Guid id) => new ParsedMatch(id, 1L);
+        public ParsedMatch CreateEmpty(Guid id) => new ParsedMatch(id, 1L, DateTimeExtensions.GetCurrentDate);
     }
 }