+++ src/history-parse/Scheduling/BaseHistoryParseJob.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Threading.Tasks;
-using CSGOStats.Services.Core.Scheduling;
+using CSGOStats.Infrastructure.Core.Scheduling;
 using CSGOStats.Services.HistoryParse.Processing;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace CSGOStats.Services.HistoryParse.Scheduling
 {
-    public abstract class BaseHistoryParseJob : JobBase
+    public abstract class BaseHistoryParseJob : BaseJob
     {
         protected abstract bool IsForcedRun { get; }