+++ src/history-parse/Scheduling/ScheduleExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using CSGOStats.Services.Core.Scheduling;
+using CSGOStats.Infrastructure.Core.Scheduling.Extensions;
 using Microsoft.Extensions.Configuration;
 using Quartz;
 
@@ -8,18 +8,14 @@ namespace CSGOStats.Services.HistoryParse.Scheduling
 {
     public static class ScheduleExtensions
     {
-        public static async Task ConfigureJobsAsync(
+        public static Task ConfigureJobsAsync(
             IScheduler scheduler,
             IServiceProvider serviceProvider,
             IConfigurationRoot configuration)
         {
-            await scheduler.ScheduleJob(
-                jobDetail: serviceProvider.CreateJobTemplate<DefaultJob>(),
-                trigger: SchedulerExtensions.CreateCronScheduledTriggerFromConfiguration(configuration, "Jobs:Default:CronExpression"));
-
-            await scheduler.ScheduleJob(
-                jobDetail: serviceProvider.CreateJobTemplate<ForcedJob>(),
-                trigger: SchedulerExtensions.CreateCronScheduledTriggerFromConfiguration(configuration, "Jobs:Forced:CronExpression"));
+            return Task.WhenAll(
+                scheduler.ScheduleJobAsync<DefaultJob>(serviceProvider, configuration),
+                scheduler.ScheduleJobAsync<ForcedJob>(serviceProvider, configuration));
         }
     }
 }